# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Biblioteca do Condomínio API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans
      security:
        - adminBearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/LoanStatus"
        - name: userId
          in: query
          schema:
            type: integer
        - name: bookId
          in: query
          schema:
            type: integer
        - name: dueSoon
          in: query
          schema:
            type: boolean
        - name: dueDateFrom
          in: query
          schema:
            type: string
            format: date
        - name: dueDateTo
          in: query
          schema:
            type: string
            format: date
        - name: reportStatus
          in: query
          schema:
            type: string
            enum: [upcoming, due_today, overdue, returned]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Loan list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoansListResponse"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Register loan
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "400":
          description: Eligibility error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan detail
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /loans/{id}/return:
    patch:
      operationId: returnLoan
      tags: [loans]
      summary: Register loan return
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ─── Reservations ─────────────────────────────────────────────────────────────
...
  /notes:
    get:
      operationId: listNotes
      tags: [notes]
      summary: List active notes
      parameters:
        - name: all
          in: query
          description: Include inactive notes (admin use)
          schema:
            type: boolean
      responses:
        "200":
          description: Notes list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotesListResponse"
    post:
      operationId: createNote
      tags: [notes]
      summary: Create note
      security:
        - adminBearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNoteRequest"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
  /notes/{id}:
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update note
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateNoteRequest"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete note
      security:
        - adminBearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # ─── Admins ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    readerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]

    # ─── Enums ─────────────────────────────────────────────────────────────────
    BookStatus:
      type: string
      enum: [draft, available, borrowed, reserved, lost, unavailable]
    UserStatus:
      type: string
      enum: [pending, active, inactive, blocked]
    LoanStatus:
      type: string
      enum: [active, returned, overdue]
...
    NoteType:
      type: string
      enum: [rule, info, announcement]

    # ─── Auth ──────────────────────────────────────────────────────────────────
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        block:
          type: string
          nullable: true
        house:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/UserStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, name, email, status, createdAt, updatedAt]
...
    BookSummary:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        author:
          type: string
        genre:
          type: string
        isbn:
          type: string
          nullable: true
        publishedYear:
          type: integer
          nullable: true
        imageUrl:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/BookStatus"
      required: [id, title, author, genre, status]
...
    LoanSummary:
      type: object
      properties:
        id:
          type: integer
        bookId:
          type: integer
        userId:
          type: integer
        status:
          $ref: "#/components/schemas/LoanStatus"
        loanDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
          nullable: true
        book:
          $ref: "#/components/schemas/BookSummary"
        user:
          $ref: "#/components/schemas/UserProfile"
      required: [id, bookId, userId, status, loanDate, dueDate]
    LoanDetail:
      allOf:
        - $ref: "#/components/schemas/LoanSummary"
        - type: object
          properties:
            createdAt:
              type: string
              format: date-time
          required: [createdAt]
    LoansListResponse:
      type: object
      properties:
        loans:
          type: array
          items:
            $ref: "#/components/schemas/LoanDetail"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required: [loans, meta]
    CreateLoanRequest:
      type: object
      properties:
        bookId:
          type: integer
        userId:
          type: integer
        dueDate:
          type: string
          format: date-time
      required: [bookId, userId]

    # ─── Reservations ──────────────────────────────────────────────────────────
...
    Note:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        type:
          $ref: "#/components/schemas/NoteType"
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, title, content, type, active, createdAt, updatedAt]
    NotesListResponse:
      type: object
      properties:
        notes:
          type: array
          items:
            $ref: "#/components/schemas/Note"
      required: [notes]
    CreateNoteRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          $ref: "#/components/schemas/NoteType"
        active:
          type: boolean
      required: [title, content]
    UpdateNoteRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        type:
          $ref: "#/components/schemas/NoteType"
        active:
          type: boolean

    # ─── Gemini ────────────────────────────────────────────────────────────────
...